requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pandas==2.2.0

# Financial Data
yfinance==0.2.54

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Sentiment Analysis
transformers==4.37.2
torch==2.2.0
vaderSentiment==3.3.2
textblob==0.17.1

# Paper Trading
alpaca-py==0.29.2
